requirements.txt
absl-py==2.0.0
accelerate==0.21.0
adal==1.2.7
addict==2.4.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiosignal==1.3.1
albumentations==1.3.1
anyio==3.7.1
appdirs==1.4.4
arel==0.2.0
asttokens==2.2.1
async-asgi-testclient==1.4.11
async-timeout==4.0.3
attrs==23.1.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.12.2
bidict==0.22.1
Bottleneck==1.3.5
brotlipy==0.7.0
cacheout==0.14.1
cachetools==4.2.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.7
colorama==0.4.6
comm==0.1.3
contourpy==1.0.5
cryptography==38.0.4
cycler==0.11.0
Cython==0.29.30
cytoolz==0.12.0
dask==2022.12.1
dataclasses==0.8
datasets==2.14.4
debugpy==1.6.7
decorator==5.1.1
diffusers==0.11.0
dill==0.3.7
distinctipy==1.2.3
dparse==0.6.3
einops==0.7.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==0.8.3
facexlib==0.3.0
fastapi==0.95.1
ffmpeg-python==0.2.0
filelock==3.8.2
filterpy==1.4.5
fonttools==4.25.0
fpie==0.2.4
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.2
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.40
giturlparse==0.12.0
gmpy2==2.1.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
grpcio==1.59.3
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.11.1
idna==3.4
imagecodecs==2021.8.26
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==6.0.1
imutils==0.5.4
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
kubernetes==26.1.0
llvmlite==0.41.1
lmdb==1.4.1
locket==1.0.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.2
multiprocess==0.70.15
munkres==1.1.4
nest-asyncio==1.5.6
networkx==3.1
numba==0.58.1
numerize==0.12
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
opencensus==0.11.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.1
pandas==1.5.2
parso==0.8.3
partd==1.4.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==23.2.1
platformdirs==3.8.0
ply==3.11
pooch==1.4.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
PTable==0.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.4
pycparser==2.21
pydantic==1.10.13
pydicom==2.4.3
Pygments==2.15.1
pynrrd==0.4.3
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.5
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qudida==0.0.4
realesrgan==0.3.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.53
safety==2.3.4
scikit-image==0.19.3
scikit-learn==1.2.0
scikit-video==1.1.11
scipy==1.10.0
seaborn==0.12.2
setuptools==68.1.2
Shapely==1.8.1
SimpleITK==2.3.1
sip==6.6.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
starlette==0.26.1
statsmodels==0.13.5
stringcase==1.2.0
supervisely==6.72.196
sympy==1.11.1
taichi==1.6.0
tb-nightly==2.16.0a20231117
tensorboard-data-server==0.7.2
tf-keras-nightly==2.16.0.dev2023111710
threadpoolctl==3.2.0
tifffile==2021.4.8
timm==0.6.13
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.3.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.28.1
trimesh==3.23.5
typing_extensions==4.6.3
tzdata==2023.3
upscalers==0.5.3
urllib3==1.26.15
uvicorn==0.24.0.post1
uvloop==0.19.0
varname==0.12.0
watchfiles==0.21.0
watchgod==0.6
wcwidth==0.2.6
websockets==10.4
Werkzeug==3.0.1
wheel==0.41.1
xxhash==2.0.2
yapf==0.40.2
yarl==1.7.2
zipp==3.15.0
